feat(resources): decouple actions to avoid inheritance
`BaseResource` is renamed to `Resource` which can now be instantiate
directly to use by passing "actions" in the constructor, so that
each resource can now have its own actions.

Lib "axios" is no longer needed because all the actions will be
defined by the user.

Tests are rewrited for the new feature.

BREAKING CHANGE: actions are moved to `.actions` and there is a
new required option `actions`.